[OWL-1730][common][nqm-mng] Code refactoring
Move some code from `common/` to `modules/nqm-mng/` as the preparation
of the following development.
