+++ modules/nqm-mng/rdb/heartbeat.go
@@ -2,8 +2,12 @@ package rdb
 
 import (
 	"github.com/Cepave/open-falcon-backend/common/db"
+	commonDb "github.com/Cepave/open-falcon-backend/common/db"
 	sqlxExt "github.com/Cepave/open-falcon-backend/common/db/sqlx"
+	gormExt "github.com/Cepave/open-falcon-backend/common/gorm"
+	nqmModel "github.com/Cepave/open-falcon-backend/common/model/nqm"
 	"github.com/Cepave/open-falcon-backend/modules/nqm-mng/model"
+	"github.com/jinzhu/gorm"
 	"github.com/jmoiron/sqlx"
 )
 
@@ -128,4 +132,128 @@ func (uoiHost *updateOrInsertHostsInTx) addHost(insertStmt *sqlx.Stmt, agent *mo
 
 func (uoiHost *updateOrInsertHostsInTx) updateHost(updateStmt *sqlx.Stmt, agent *model.AgentHeartbeat) {
 	uoiHost.result.RowsAffected += updateAndGetRowsAffected(updateStmt, agent)
+}
+
+type UpdateNqmAgentProcessor struct {
+	Reqs []*model.NqmAgentHeartbeatRequest
+}
+
+func (p *UpdateNqmAgentProcessor) InTx(tx *sqlx.Tx) commonDb.TxFinale {
+	updateStmt := sqlxExt.ToTxExt(tx).Preparex(`
+	UPDATE nqm_agent
+	SET ag_hostname = ?,
+		ag_ip_address = ?,
+		ag_last_heartbeat = FROM_UNIXTIME(?)
+	WHERE ag_connection_id = ?
+		AND ag_last_heartbeat < FROM_UNIXTIME(?)
+	`)
+
+	for _, e := range p.Reqs {
+		updateStmt.MustExec(
+			e.Hostname,
+			e.IpAddress,
+			e.Timestamp,
+			e.ConnectionId,
+			e.Timestamp,
+		)
+	}
+	return commonDb.TxCommit
+}
+
+func SelectByConnId(connId string) *nqmModel.Agent {
+	var selectAgent = DbFacade.GormDb.Model(&nqmModel.Agent{}).
+		Select(`
+			ag_id, ag_name, ag_connection_id, ag_hostname, ag_ip_address, ag_status, ag_comment, ag_last_heartbeat,
+			COUNT(DISTINCT pt.pt_id) AS ag_num_of_enabled_pingtasks,
+			isp_id, isp_name, pv_id, pv_name, ct_id, ct_name, nt_id, nt_value,
+			GROUP_CONCAT(gt.gt_id ORDER BY gt_name ASC SEPARATOR ',') AS gt_ids,
+			GROUP_CONCAT(gt.gt_name ORDER BY gt_name ASC SEPARATOR '\0') AS gt_names
+		`).
+		Joins(`
+			LEFT JOIN
+			nqm_agent_ping_task AS apt
+			ON ag_id = apt.apt_ag_id
+			LEFT JOIN
+			nqm_ping_task AS pt
+			ON apt.apt_pt_id = pt.pt_id AND pt.pt_enable=true
+			INNER JOIN
+			owl_isp AS isp
+			ON ag_isp_id = isp.isp_id
+			INNER JOIN
+			owl_province AS pv
+			ON ag_pv_id = pv.pv_id
+			INNER JOIN
+			owl_city AS ct
+			ON ag_ct_id = ct.ct_id
+			INNER JOIN
+			owl_name_tag AS nt
+			ON ag_nt_id = nt.nt_id
+			LEFT OUTER JOIN
+			nqm_agent_group_tag AS agt
+			ON ag_id = agt.agt_ag_id
+			LEFT OUTER JOIN
+			owl_group_tag AS gt
+			ON agt.agt_gt_id = gt.gt_id
+		`).
+		Where("ag_connection_id = ?", connId).
+		Group(`
+			ag_id, ag_name, ag_connection_id, ag_hostname, ag_ip_address, ag_status, ag_comment, ag_last_heartbeat,
+			isp_id, isp_name, pv_id, pv_name, ct_id, ct_name, nt_id, nt_value
+		`)
+
+	var loadedAgent = &nqmModel.Agent{}
+	selectAgent = selectAgent.Find(loadedAgent)
+
+	if selectAgent.Error == gorm.ErrRecordNotFound {
+		return nil
+	}
+	gormExt.ToDefaultGormDbExt(selectAgent).PanicIfError()
+
+	loadedAgent.AfterLoad()
+
+	if !loadedAgent.Status {
+		return nil
+	}
+
+	return loadedAgent
+}
+
+func NotNew(r *model.NqmAgentHeartbeatRequest) bool {
+	return DbFacade.SqlxDbCtrl.GetOrNoRow(
+		new(int),
+		`
+			SELECT ag_id
+			FROM nqm_agent
+			WHERE ag_connection_id = ?
+		`,
+		r.ConnectionId,
+	)
+}
+
+func Insert(r *model.NqmAgentHeartbeatRequest) {
+	DbFacade.SqlxDb.MustExec(`
+		INSERT INTO host(hostname, ip, agent_version, plugin_version)
+		VALUES(?, ?, '', '')
+		ON DUPLICATE KEY UPDATE
+			ip = VALUES(ip)
+	`,
+		r.Hostname,
+		string(r.IpAddress),
+	)
+	DbFacade.SqlxDb.MustExec(`
+		INSERT INTO nqm_agent(ag_connection_id, ag_hostname, ag_ip_address, ag_last_heartbeat, ag_hs_id)
+		SELECT ?, ?, ?, FROM_UNIXTIME(?), id
+		FROM host
+		WHERE hostname = ?
+		ON DUPLICATE KEY UPDATE
+			ag_hostname = VALUES(ag_hostname),
+			ag_ip_address = VALUES(ag_ip_address),
+			ag_last_heartbeat = VALUES(ag_last_heartbeat)
+	`,
+		r.ConnectionId,
+		r.Hostname,
+		r.IpAddress,
+		r.Timestamp,
+		r.Hostname,
+	)
 }
