+++ modules/nqm-mng/restful/package.go
@@ -1,16 +1,10 @@
 package restful
 
 import (
-	commonDb "github.com/Cepave/open-falcon-backend/common/db"
-	commonNqmDb "github.com/Cepave/open-falcon-backend/common/db/nqm"
-	sqlxExt "github.com/Cepave/open-falcon-backend/common/db/sqlx"
 	commonGin "github.com/Cepave/open-falcon-backend/common/gin"
 	"github.com/Cepave/open-falcon-backend/common/gin/mvc"
 	log "github.com/Cepave/open-falcon-backend/common/logruslog"
-	nqmModel "github.com/Cepave/open-falcon-backend/common/model/nqm"
-	commonQueue "github.com/Cepave/open-falcon-backend/common/queue"
-	"github.com/jmoiron/sqlx"
-	gin "gopkg.in/gin-gonic/gin.v1"
+	"gopkg.in/gin-gonic/gin.v1"
 )
 
 var logger = log.NewDefaultLogger("INFO")
@@ -42,14 +36,6 @@ var cacheConfig *CacheConfig = &CacheConfig{}
 
 func InitCache(config *CacheConfig) {
 	cacheConfig = config
-}
-
-var heartbeatConfig *commonQueue.Config
-
-func InitHeartbeat(config *commonQueue.Config) {
-	heartbeatConfig = config
-	commonNqmDb.HeartbeatReqQueue = commonQueue.New()
-	go drain()
 }
 
 func initApi() {
@@ -98,40 +84,4 @@ func initApi() {
 	v1.POST("/agent/heartbeat", mvcBuilder.BuildHandler(agentHeartbeat))
 
 	router.GET("/health", health)
-}
-
-type updateNqmAgentProcessor struct {
-	reqs []*nqmModel.AgentHeartbeatRequest
-}
-
-func (p *updateNqmAgentProcessor) InTx(tx *sqlx.Tx) commonDb.TxFinale {
-	updateStmt := sqlxExt.ToTxExt(tx).Preparex(`
-	UPDATE nqm_agent
-	SET ag_hostname = ?,
-		ag_ip_address = ?,
-		ag_last_heartbeat = FROM_UNIXTIME(?)
-	WHERE ag_connection_id = ?
-		AND ag_last_heartbeat < FROM_UNIXTIME(?)
-	`)
-
-	for _, e := range p.reqs {
-		updateStmt.MustExec(
-			e.Hostname,
-			e.IpAddress,
-			e.Timestamp,
-			e.ConnectionId,
-			e.Timestamp,
-		)
-	}
-	return commonDb.TxCommit
-}
-
-func drain() {
-	for {
-		reqs := commonNqmDb.HeartbeatReqQueue.DrainNWithDurationByType(heartbeatConfig, new(nqmModel.AgentHeartbeatRequest)).([]*nqmModel.AgentHeartbeatRequest)
-		updateTx := &updateNqmAgentProcessor{
-			reqs: reqs,
-		}
-		commonNqmDb.DbFacade.SqlxDbCtrl.InTx(updateTx)
-	}
 }
